Accept explicit kind signatures in type variables
PolyKinds extension annotates unused type variables with a polymorphic
kind variable; this triggers a failure when running `getTyConInfo`, as
those variables cease to be recognized as _plain_.

The user can avoid this by explicitly annotating the variable with kind `*`.

`getTyConInfo` is changed to treat as plain type variables those with explicit
kind `*`.